consumer1/requirements.txt
kafka-python
pandas
minio
